requirements.txt
google-generativeai
python-dotenv
beautifulsoup4
requests
pandas
tqdm
anthropic
langchain
langchain-anthropic
google-generativeai